Engines page now using auto-generated model class.
